fix: update validation to zod, and refractoring all usecase

+++ src/user/user.service.ts
@@ -4,12 +4,14 @@ import {
   NotFoundException,
 } from '@nestjs/common';
 import { FileVisibility } from '@prisma/client';
-import { User as UserModel } from 'src/models';
-import { PrismaService } from 'src/prisma/prisma.service';
-import { UpdateUserDto } from './dto';
+import { UserModel } from '@src/models';
+import { PrismaService } from '@src/prisma/prisma.service';
 import * as bcrypt from 'bcrypt';
-import { FileStorageOptions } from 'src/file-storage/types';
-import { deleteFile, uploadFile } from 'src/common/utils/file-storage';
+import { FileStorageOptions } from '@src/file-storage/types';
+import { deleteFile, uploadFile } from '@src/common/utils/file-storage';
+import { ValidationService } from '@src/common/validation.service';
+import { UserValidation } from './user.validation';
+import { UpdateUserRequest } from './dto';
 
 @Injectable()
 export class UserService {
@@ -18,7 +20,10 @@ export class UserService {
     visibility: this.fileVisibility,
     storageFolder: 'profiles',
   };
-  constructor(private readonly prismaService: PrismaService) {}
+  constructor(
+    private readonly prismaService: PrismaService,
+    private validationService: ValidationService,
+  ) {}
 
   async get(user_id: string): Promise<UserModel> {
     const user = await this.prismaService.user.findUnique({
@@ -35,9 +40,15 @@ export class UserService {
 
   async update(
     user_id: string,
-    data: UpdateUserDto,
+    data: UpdateUserRequest,
     file?: Express.Multer.File,
   ): Promise<UserModel> {
+    const UpdateUserRequest = this.validationService.validate(
+      UserValidation.UPDATE_USER_REQUEST,
+      data,
+    );
+    console.log('UpdateUserRequest', UpdateUserRequest);
+
     const user = await this.prismaService.user.findUnique({
       where: {
         user_id,
@@ -48,17 +59,23 @@ export class UserService {
       throw new NotFoundException('User not found');
     }
 
-    if (data.password) {
-      data.password = await this.hashPassword(data.password, user.salt);
+    if (UpdateUserRequest.password) {
+      UpdateUserRequest.password = await this.hashPassword(
+        UpdateUserRequest.password,
+        user.salt,
+      );
     }
     try {
       const transaction = await this.prismaService.$transaction(
         async (prisma) => {
-          if (data.profile !== user.profile && file) {
-            data.profile = await uploadFile(file, this.fileStorageOptions);
+          if (file && file.originalname !== user.profile) {
+            UpdateUserRequest.profile = await uploadFile(
+              file,
+              this.fileStorageOptions,
+            );
           }
 
-          if (user.profile) {
+          if (file && file.originalname !== user.profile) {
             await deleteFile(user.profile, this.fileStorageOptions);
           }
 
@@ -66,7 +83,7 @@ export class UserService {
             where: {
               user_id: user.user_id,
             },
-            data,
+            data: UpdateUserRequest,
           });
         },
       );
